the code is written a story done a work complete skills can be edited
